🔒 Warden: Fix integer overflow vulnerability in prepare_scanline (#709)
- Patched `src/rasterizer/core.rs` to fix an integer overflow panic (DoS vector)
- The expression `-xs as f32` evaluated the negation `(-xs)` before the cast
- If `xs == i32::MIN`, this caused an overflow panic or wrap-around
- Changed to `-i64::from(xs) as f32` to resolve it securely
- Added a targeted proptest/unit test to verify the fix
- Logged vulnerability discovery in `.jules/warden.md`

+++ src/rasterizer/core.rs
@@ -60,7 +60,7 @@ pub(crate) fn prepare_scanline<'a>(
     let mut xe = x_end;
 
     if xs < 0 {
-        let diff = -xs as f32;
+        let diff = -i64::from(xs) as f32;
         z += diff * dz_dx;
         xs = 0;
     }
